requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==6.0.0
bootstrap4==0.1.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
comm==0.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
dataclasses-json==0.5.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.2
django-admin-interface==0.19.2
django-admin-volt==1.0.4
django-bootstrap4==22.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.21.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.4.2
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.3
inflection==0.5.1
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
isoduration==20.11.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
langchain==0.0.158
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mistune==2.0.4
mongoengine==0.25.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-codec==1.3.2
openapi-schema-pydantic==1.2.4
packaging==22.0
pandocfilters==1.5.0
parso==0.8.3
phonenumbers==8.13.11
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
PyJWT==1.7.1
pymongo==3.12.1
pypdf==3.8.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Send2Trash==1.8.0
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==2.0.12
sqlparse==0.2.4
stack-data==0.6.2
stripe==5.2.0
tenacity==8.2.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2021.5
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.13.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
whitenoise==6.2.0
yarl==1.9.2
